Moved SortSelector to new _molecules-folder
